+++ proxysiphon/__init__.py
@@ -1,3 +1,3 @@
-from proxysiphon.records import read_ncdc, NcdcRecord
+from proxysiphon.records import read_ncdc, read_lgm, read_petm
 from proxysiphon.agemodel import get_deltar_online, fit_agedepthmodel, date_proxy
 from . import qcreport